move installation docsgen to a new file
